Get to "lists" page
This Selenium implementation can login via Google and navigate to the Medium "lists" page, so the next step is copying the relevant information and removing the highlights from the list.

Unfortunately this cannot be implemented quite yet because the "Unhighlight" button is broken. I reached out to the Medium team and they are aware of the issue and are working to get it fixed. As soon as that is done, I can finish out the program. Before then I may push some skeleton code that will iterate over the highlights in preparation for when that functionality is working again.